pre-commit: Manage vale with pre-commit
This replaces the previous script hook with two python hooks reusing the
same dependency.
The first to run a single instance of vale sync, the other to let
pre-commit run multiple instances of vale run.

This simplifies the workflow for contributors as they do not need to
install vale seperately, or install all of the dependencies to commit.