ci: Enforce CISO / DPO controls
As discussed in GOTO public docs, we want to make it harder for people to
forget to update CISO / DPO controls. This PR proposes a way to ensure that.

`kentaro-m/task-completed-checker-action@v0.1.2` seems to be the most
popular GitHub Action to enforce checkboxes being ticked when a PR is
submitted.

Fixes #1203