chore: Tech radar fixes (#1216)
* Add our own analytics

Unfortunately, we missed that the tech radar included upstream analytics via
Google Analytics. Removed it and replace it with our privacy-friendly analytics.

* Enforce restrictive Content-Security-Policy

Avoids this website "talking to strangers", both now and in the future.

* Vendor in D3 and upgrade to version 7

This is [ISC licensed](https://github.com/d3/d3/blob/main/LICENSE), so it's okay
to do so.

* fix: Argo CD is ADOPT